#1 change precision in MathUtils

+++ RtCs.MathUtils/Geometry/Line3D.cs
@@ -17,33 +17,33 @@ namespace RtCs.MathUtils.Geometry
         public Vector3 Point1
         { get; set; }
 
-        public Vector3 Along(double inRatio)
+        public Vector3 Along(float inRatio)
             => Point0 + ((Point1 - Point0) * inRatio);
 
         public Vector3 Vector => Point1 - Point0;
         public Vector3 Direction => Vector.Normalized;        
-        public double Length => Vector.Length;
+        public float Length => Vector.Length;
 
         public Vector3 Project(Vector3 inPoint)
-            => Project(inPoint, out double _);
+            => Project(inPoint, out float _);
 
-        public Vector3 Project(Vector3 inPoint, out double outParameter)
+        public Vector3 Project(Vector3 inPoint, out float outParameter)
         {
             if (Vector.IsZero) {
-                outParameter = 0.0;
+                outParameter = 0.0f;
                 return Point0;
             }
             outParameter = Vector3.Dot(Direction, inPoint - Point0) / Length;
             return Point0 + (outParameter * Vector);
         }
 
-        public double DistanceTo(Vector3 inPoint)
+        public float DistanceTo(Vector3 inPoint)
         {
-            Project(inPoint, out double t);
-            return (inPoint - Along(t.Clamp(0.0, 1.0))).Length;
+            Project(inPoint, out float t);
+            return (inPoint - Along(t.Clamp(0.0f, 1.0f))).Length;
         }
 
-        public bool IsIntersect(Plane inPlane, out double outT, out Vector3 outPosition)
+        public bool IsIntersect(Plane inPlane, out float outT, out Vector3 outPosition)
         {
             // refer http://www.sousakuba.com/Programming/gs_plane_line_intersect.html
 
@@ -53,20 +53,20 @@ namespace RtCs.MathUtils.Geometry
                 return false;
             }
 
-            double n_c0 = Vector3.Dot(inPlane.Normal, Point0 - inPlane.Point);
-            double n_c1 = Vector3.Dot(inPlane.Normal, Point1 - inPlane.Point);
-            double abs_n_c0 = Math.Abs(n_c0);
-            double abs_n_c1 = Math.Abs(n_c1);
+            float n_c0 = Vector3.Dot(inPlane.Normal, Point0 - inPlane.Point);
+            float n_c1 = Vector3.Dot(inPlane.Normal, Point1 - inPlane.Point);
+            float abs_n_c0 = Math.Abs(n_c0);
+            float abs_n_c1 = Math.Abs(n_c1);
 
             if (abs_n_c0.AlmostZero()) {
                 // Point0 on plane
-                outT = 0.0;
+                outT = 0.0f;
                 outPosition = Point0;
                 return true;
             }
             if (abs_n_c1.AlmostZero()) {
                 // Point1 on Plane
-                outT = 1.0;
+                outT = 1.0f;
                 outPosition = Point1;
                 return true;
             }
@@ -82,37 +82,37 @@ namespace RtCs.MathUtils.Geometry
             }
 
             outPosition = Point0 + (Vector * outT);
-            return outT.InRange(0.0, 1.0);
+            return outT.InRange(0.0f, 1.0f);
         }
 
         public bool IsIntersect(AABB3D inAABB)
-            => IsIntersect(inAABB, out double _, out double _);
+            => IsIntersect(inAABB, out float _, out float _);
 
-        public bool IsIntersect(AABB3D inAABB, out double outParamNear, out double outParamFar)
+        public bool IsIntersect(AABB3D inAABB, out float outParamNear, out float outParamFar)
         {
             // refer http://marupeke296.com/COL_3D_No18_LineAndAABB.html
 
-            outParamNear = double.NaN;
-            outParamFar = double.NaN;
+            outParamNear = float.NaN;
+            outParamFar = float.NaN;
 
             Vector3 p = Point0;
             Vector3 d = Direction;
             Vector3 min = inAABB.Min;
             Vector3 max = inAABB.Max;
 
-            double near = double.MinValue;
-            double far = double.MaxValue;
+            float near = float.MinValue;
+            float far = float.MaxValue;
             for (int i = 0; i < 3; ++i) {
                 if (!d[i].AlmostZero()) {
                     if ((p[i] < min[i]) || (max[i] < p[i])) {
                         return false;
                     }
 
-                    double odd = 1.0 / d[i];
-                    double t0 = (min[i] - p[i]) * odd;
-                    double t1 = (max[i] - p[i]) * odd;
+                    float odd = 1.0f / d[i];
+                    float t0 = (min[i] - p[i]) * odd;
+                    float t1 = (max[i] - p[i]) * odd;
                     if (t0 > t1) {
-                        double tmp = t0;
+                        float tmp = t0;
                         t0 = t1;
                         t1 = tmp;
                     }
@@ -148,8 +148,8 @@ namespace RtCs.MathUtils.Geometry
         { get; set; } = null;
         public Vector3 Position
         { get; set; } = new Vector3();
-        public double LineParameter
-        { get; set; } = double.NaN;
+        public float LineParameter
+        { get; set; } = float.NaN;
         public Vector3? Normal
         { get; set; } = null;
     }
